move epsilon.jl test to deterministic, clean deterministic output test
