+++ test/deterministic/deterministic_ouput_t1_t5.jl
@@ -1,7 +1,6 @@
 @testset "time 1 and 5 deterministic" begin
 
     dir = @__DIR__
-
     T = 1
     parameters = BeforeIT.AUSTRIA2010Q1.parameters
     initial_conditions = BeforeIT.AUSTRIA2010Q1.initial_conditions
@@ -28,18 +27,17 @@
         end
     end
 
-
-    T = 4
-    parameters = matread(joinpath(dir, "../../data/austria/parameters/2010Q1.mat"))
-    initial_conditions = matread(joinpath(dir, "../../data/austria/initial_conditions/2010Q1.mat"))
+    T = 5
+    parameters = BeforeIT.AUSTRIA2010Q1.parameters
+    initial_conditions = BeforeIT.AUSTRIA2010Q1.initial_conditions
     model = BeforeIT.initialise_model(parameters, initial_conditions, T)
     data = BeforeIT.initialise_data(model)
-    for t in 1:T-1
+    for t in 1:T
         BeforeIT.one_epoch!(model; multi_threading = false)
         BeforeIT.update_data!(data, model)
     end
 
-    output_t5 = matread(joinpath(dir, "../matlab_code/output_t5.mat"))list
+    output_t5 = matread(joinpath(dir, "../matlab_code/output_t5.mat"))
 
     # confront results between julia and matlab code
 
@@ -47,10 +45,10 @@
         julia_output = getfield(data, fieldname)
         matlab_output = output_t5[string(fieldname)]
 
-        if length(size(julia_output)) == 1
-            @test isapprox(julia_output, matlab_output', rtol = 1e-4)
+        if length(julia_output) == 6
+            @test isapprox(julia_output[2:end], matlab_output', rtol = 1e-4)
         else
-            @test isapprox(julia_output, matlab_output, rtol = 1e-5)
+            @test isapprox(julia_output[2:end, :], matlab_output, rtol = 1e-5)
         end
     end
