+++ test/deterministic/runtests_deterministic.jl
@@ -1,4 +1,5 @@
 @warn "making the model deterministic all subsequent calls to the model will be deterministic."
+include("epsilon.jl")
 include("make_model_deterministic.jl")
 include("initialize_deterministic.jl")
 include("one_epoch_deterministic.jl")
