+++ test/runtests.jl
@@ -1,11 +1,9 @@
 using BeforeIT, Test
 
 #utils
-include("utils/epsilon.jl")
 include("utils/positive.jl")
 include("utils/randpl.jl")
 include("utils/nfvar3_and_estimate.jl")
-
 include("utils/estimations.jl")
 
 # search_and_matching
